+++ src/Phetch.Core/Query.cs
@@ -476,23 +476,6 @@ public class Query<TArg, TResult> : IQuery<TArg, TResult>
 }
 
 /// <summary>
-/// An alternate version of <see cref="Query{TArg, TResult}"/> for queries with no return value.
-/// </summary>
-/// <remarks>Aside from having no return value, this functions identically to a normal Query</remarks>
-[Obsolete("Use Query<TArg, Unit> instead, which is equivalent.", true)]
-[ExcludeFromCodeCoverage]
-public class Mutation<TArg> : Query<TArg, Unit>
-{
-    internal Mutation(
-        QueryCache<TArg, Unit> cache,
-        QueryOptions<TArg, Unit>? options,
-        EndpointOptions<TArg, Unit> endpointOptions
-    ) : base(cache, options, endpointOptions)
-    {
-    }
-}
-
-/// <summary>
 /// An alternate version of <see cref="Query{TArg, TResult}"/> for queries with no parameters.
 /// </summary>
 /// <remarks>Aside from having no parameters, this functions identically to a normal Query</remarks>